# docs/research/main_specs/1.0.0-mvp/api/openapi.yaml
...
openapi: 3.1.0
info:
  title: Helix OTA Control Plane API
  version: 1.0.0-MVP
  summary: REST API for the Helix OTA modular-monolith control plane (Android 15 first).
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  description: >
    Control-plane endpoints (auth, devices, artifacts, releases, deployments, client
    update-check + telemetry). JSON responses negotiate Brotli/gzip; the referenced
    artifact-byte path is served identity-encoded, ZIP_STORED, with HTTP Range support.
    Response-header conventions (applied to every negotiated JSON response, not repeated per
    operation): 'Vary: Accept-Encoding' is always set (the Vary header component), and 'X-Request-Id'
    is always returned. Rate-limited routes additionally return RateLimit-Limit / RateLimit-Remaining
    / RateLimit-Reset (draft IETF RateLimit fields); on 429 a Retry-After header is returned.
  contact:
    name: Helix OTA control-plane team
  # license: intentionally omitted — project license is UNVERIFIED in this revision.
servers:
  - url: https://{host}/api/v1
    description: Helix OTA control plane (HTTP/3 primary, HTTP/2 fallback).
    variables:
      host:
        default: helix.example
...
security:
  - bearerAuth: []
paths:
...
  /deployments/{deploymentId}:
    get:
      tags: [deployments]
      operationId: getDeployment
      summary: Read a deployment with aggregate progress.
      parameters:
        - $ref: '#/components/parameters/DeploymentId'
      responses:
        '200':
          description: Deployment status.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DeploymentStatus' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /webhooks:
    post:
      tags: [webhooks]
      operationId: createWebhook
      summary: Register a webhook for event-driven delivery.
      description: >
        RBAC operator/admin. Creates a project-scoped webhook callback
        (handlers_webhook.go:handleCreateWebhook).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WebhookCreate' }
      responses:
        '201':
          description: Webhook registered.
          headers:
            X-Request-Id: { $ref: '#/components/headers/XRequestId' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WebhookView' }
        '400': { $ref: '#/components/responses/ValidationFailed' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
    get:
      tags: [webhooks]
      operationId: listWebhooks
      summary: List webhooks (account-scoped).
      description: RBAC viewer+. Returns webhooks for the token's account (handlers_webhook.go:handleListWebhooks).
      responses:
        '200':
          description: Webhooks list.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WebhookList' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /webhooks/{id}:
    delete:
      tags: [webhooks]
      operationId: deleteWebhook
      summary: Delete a webhook.
      description: RBAC operator/admin (handlers_webhook.go:handleDeleteWebhook).
      parameters:
        - $ref: '#/components/parameters/WebhookId'
      responses:
        '204':
          description: Webhook deleted.
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

  # --- Branches (project-scoped release channels) ---
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        OAuth2/JWT access token (admin/operator/viewer) or device-scoped token (role=device).
        RBAC enforced per-route; device tokens may act only on their own deviceId.
    provisioningToken:
      type: http
      scheme: bearer
      description: >
        Alternative auth for POST /devices/register only: a provisioning token that authorizes a
        device to enrol itself when no operator bearer is available. The global bearerAuth (operator
        role) is the primary path; this scheme is the alternative (either satisfies the operation).
  headers:
    XRequestId:
      description: Correlation id; mirrors error.request_id.
      schema: { type: string }
...
  parameters:
...
    DeploymentId:
      name: deploymentId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    WebhookId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    ValidationFailed:
      description: Request failed validation.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthenticated:
      description: Missing, invalid, or expired credentials.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Authenticated but lacks the required role or resource ownership.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
...
    Error:
      type: object
      required: [error]
      additionalProperties: false
      properties:
        error:
          type: object
          required: [code, message]
          additionalProperties: false
          properties:
            code:
              type: string
              enum:
                - UNAUTHENTICATED
                - FORBIDDEN
                - NOT_FOUND
                - VALIDATION_FAILED
                - CONFLICT
                - UNSUPPORTED_MEDIA_TYPE
                - PAYLOAD_TOO_LARGE
                - RATE_LIMITED
                - SIGNATURE_INVALID
                - HASH_MISMATCH
                - VERSION_NOT_MONOTONIC
                - INTERNAL
            message: { type: string }
            request_id: { type: string }
            details:
              type: array
              items:
                type: object
                additionalProperties: false
                properties:
                  field: { type: string }
                  issue: { type: string }
...
    Deployment:
      type: object
      required: [deployment_id, release_id, strategy, status, created_at]
      additionalProperties: false
      properties:
        deployment_id: { type: string, format: uuid }
        release_id: { type: string, format: uuid }
        strategy: { type: string, const: all-targets }
        group: { type: string }
        status: { type: string, enum: [active, completed, halted] }
        target_count: { type: integer, minimum: 0 }
        created_at: { type: string, format: date-time }
    DeploymentStatus:
      allOf:
        - { $ref: '#/components/schemas/Deployment' }
        - type: object
          additionalProperties: false
          properties:
            progress:
              type: object
              additionalProperties: false
              properties:
                pending: { type: integer, minimum: 0 }
                downloading: { type: integer, minimum: 0 }
                installed: { type: integer, minimum: 0 }
                succeeded: { type: integer, minimum: 0 }
                failed: { type: integer, minimum: 0 }
...
    WebhookCreate:
      type: object
      required: [url, events]
      additionalProperties: false
      properties:
        url: { type: string, format: uri }
        events:
          type: array
          minItems: 1
          items:
            type: string
            enum: [artifact.uploaded, release.published, deployment.created, deployment.completed, rollout.phase_advanced, device.registered]
        project_id: { type: string, format: uuid }
        secret: { type: string, description: HMAC shared secret for webhook payload signing. }
        active: { type: boolean, default: true }
    WebhookView:
      type: object
      required: [id, url, events, active, created_at]
      additionalProperties: false
      properties:
        id: { type: string, format: uuid }
        url: { type: string, format: uri }
        events:
          type: array
          items: { type: string }
        project_id: { type: string, format: uuid }
        active: { type: boolean }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    WebhookList:
      type: object
      required: [items]
      additionalProperties: false
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/WebhookView' }

    # --- branches (handlers_branches.go, migration 5) ---
...